Make proto changes to support GetTC by ID from file system
The current implementation of GetTC is only capable of querying the
database to return a unique testcase. This PR adds the model changes
required to support querying for a single testcase from the file system.
The actual implementation would be covered in later PRs to avoid noise
due to generated code.

Signed-off-by: Aerowisca <aerowisca@gmail.com>