Add feature to pipe or redirect the logs via user flags
The current implementation prints the entire logs in the console. This
is not effective when integrated with other applications. To give the
user complete flexibility, this commit adds support for a flag using
which the user can ask the logs to either be printed on the console, or
to a file of their choice (via another flag), or pipe the logs to both
console and file.

Signed-off-by: Aerowisca <aerowisca@gmail.com>