Refactored TestReplaceHttpFields to match keploy way
Here I've refactored so it becomes more alike the way keploy
do unit tests (table-driven pattern). Its important to mention
this commits solve the problem of pointing to the same object
from the last. Another observation is that changes will need
to be made on the mock.go file as tests are not passing. It
seems to be a bug related to the domain case at the switch.

Signed-off-by: cassiozareck <cassiomilczareck@gmail.com>