Refactor pipe logs and creating a centralized output interface
Now that we have two ways of outputs its not good for individual modules
verify if it needs to export or to print in terminal. Instead, one struct
should handle that. The server.Output embed the logExport object as well
the zap. This means we can use normal zap operations and it will still
be redirected to file if the log path was defined as the zap will implement
the logExport struct.
Use an centralized class for outputs has the advantage of reusability,
scalability and less refactoring if by somehow we decide zap isnt going
to be used anymore in Keploy. In latter commits its good to update some
functions to accept this general Struct so we dont need to to always
return the inner zap logger inside the server.Output struct and instead
make use of server.Output methods that calls zap below it. But for now
there's no need for it.
Also its important to mention I've removed the
if conf.EnableDebugger {
		logConf.Level.SetLevel(zap.DebugLevel)
	}
line because in fact it was not having impact on anything as the logger
was already built, latter commits I'll see how to implement it more
properly.

Signed-off-by: cassiozareck <cassiomilczareck@gmail.com>