Add port flag and remove pid flag
PID of a process listening on a port is ephemeral and can change a lot
(depending on whether the binary was restarted). We should instead
provide the user with a flag for the port on which their application is
running. We can do the heavy lifting of figuring out the pid of that
process by ourselves, using the os/exec package.

Signed-off-by: Aerowisca <aerowisca@gmail.com>