fix(serve): runs the proxy server after loading ebpf hooks
since the keploy pid is extracted from the user space progrom pid and passed into the ebpf map from
the c code itself when sys-bind is called. And sys-bind is called when a port allocation occurs so,
need to start proxy server when the fetchTestsets is fired to load ebpf hooks correctly into the
kernel.

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>