fix: redirect dns calls by configuring nsswitch config file in linux (#1678)
* fix: redirect dns calls by configuring nsswitch config file in linux

linux distros like fedora uses systemd-resolve for dns resolution which makes unix socket with the application due to which keploy is not able to redirect them to proxy in test mode. This results in test failure. This is handled by editing the nsswitch.conf of /etc in linux to not use resolve in test mode

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

---------

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>
Former-commit-id: 7fe9bc0846876c7ddf861bf08d6a726baf2091f8