fix(cli): adds config as a field in cmdConfigurator and serviceProvider  (#1718)
* fix(cli): adds config as a field in cmdConfigurator and serviceProvider

the config is present in the function signature due to which different implementations of the
interface is restricted to use the oss config type. This makes a challenge in the enterprise to add
its own config type for enterprise flags. Thats why the config is now a member field of the struct
that implements cmdConfigurator and serviceProvider interfaces.

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* style: removes linting errors

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

---------

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>
Former-commit-id: e35909158d3728b25d58304be55ba1ff7c7e2972