fix(mongo): passes the password in the function arguments (#1729)
* fix(cli): adds config as a field in cmdConfigurator and serviceProvider

the config is present in the function signature due to which different implementations of the
interface is restricted to use the oss config type. This makes a challenge in the enterprise to add
its own config type for enterprise flags. Thats why the config is now a member field of the struct
that implements cmdConfigurator and serviceProvider interfaces.

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* fix(mongo): passes the password in arguments instead of global variable

the password global variable in the integrations/mongo results in race conditions. To fix this now,
the password is passed in the function arguments

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* style: removes commented code from mongo parser

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* fix(mongo): uses sync map for authMessage global variable to prevent race conditions

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* style(mongo): removes commented code

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

* style: fixes linting issues

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>

---------

Signed-off-by: re-Tick <jain.ritik.1001@gmail.com>