Replay: fix preprocess only when command type is native
we can't preprocess in case of commandtype as docker because preprocess appends language specific coverage executable to the command provided and the command provided by the user in case of docker would start with "docker", so we don't want to append it in provided command rather we want to append it in the command to  run in the docker container which can be extracted through docker client. but this approach is not good as after modifying the docker internal command we need to restart the container, that's why we are taking a different approach in which users have to edit their application command in dockerfile so preprocess won't happen from our end.

Signed-off-by: Akash Kumar <meakash7902@gmail.com>