postgres-parser: fixed the preparedStatementMatch logic to consider packet types which doesn't pass checkIfPs function
the correct mock should be able to match in preparedStatementMatch, but it is not because the pgrequest packet headers is not of the form B E B E B E ..... , basically, we are checking if Bind(B) is trying to refer to a previously cached query (by referring it with an identifier), In other words, if Bind is not preceeded by a Parse(P), if B is not preceeded with a P then we can conclude that the B is trying to refer to an identifier which was previously cached in the previous pg request. that's the reason I believe what checkIfPS is trying to validate.

but in preparedStatementMatch main for loop (match) logic, we are only letting pgrequests with PacketTypes which has all Bind statements trying to refer to a previously cached request (B E B E B E),

But technically, if a single bind statement is trying to refer to the previously cached identifier we should let the pgrequests pass (move to the match logic loop), Like B E P B E should not be stopped at the gates of PreparedStatementMatch.

Signed-off-by: Akash Kumar <meakash7902@gmail.com>