fix: wrong mock matching which results in testcase failure (#2678)
* fix: wrong mock matching

Before each testcase, we fetch the copy of mocks in our DS, due to this we are taking into account that some mocks may be already used, and this creates problem.

Current mock matching flow, is to first filter out mocks based on req-resp timestamp, so if two requests are made asynchronously then the same set of mocks would be matched. Since after each testcase, we are not taking into account the consumed mocks, we may get a match to a mock which has already been used in a previous testcase, this may lead to failed testcase.
To solve this, I used the already existing totalConsumedMocks set to filter out unused mocks before matching algorithm.

Signed-off-by: Akash Kumar <meakash7902@gmail.com>

* postgres-parser: fixed the preparedStatementMatch logic to consider packet types which doesn't pass checkIfPs function

the correct mock should be able to match in preparedStatementMatch, but it is not because the pgrequest packet headers is not of the form B E B E B E ..... , basically, we are checking if Bind(B) is trying to refer to a previously cached query (by referring it with an identifier), In other words, if Bind is not preceeded by a Parse(P), if B is not preceeded with a P then we can conclude that the B is trying to refer to an identifier which was previously cached in the previous pg request. that's the reason I believe what checkIfPS is trying to validate.

but in preparedStatementMatch main for loop (match) logic, we are only letting pgrequests with PacketTypes which has all Bind statements trying to refer to a previously cached request (B E B E B E),

But technically, if a single bind statement is trying to refer to the previously cached identifier we should let the pgrequests pass (move to the match logic loop), Like B E P B E should not be stopped at the gates of PreparedStatementMatch.

Signed-off-by: Akash Kumar <meakash7902@gmail.com>
---------

Signed-off-by: Akash Kumar <meakash7902@gmail.com>
Co-authored-by: Gourav kumar <44055698+gouravkrosx@users.noreply.github.com>
Co-authored-by: Yash Khare <khareyash05@gmail.com>